Move components into seperate files
