+++ apps/contentful-apps/components/content-import/utils.ts
@@ -1,0 +1,9 @@
+export type FileData = string[][]
+
+export const isReferenceField = (field: { type: string }) => {
+  return !(
+    field.type === 'Symbol' ||
+    field.type === 'Integer' ||
+    field.type === 'RichText'
+  )
+}
