+++ apps/contentful-apps/next.config.js
@@ -1,5 +1,8 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const withNx = require('@nx/next/plugins/with-nx')
+const { composePlugins, withNx } = require('@nx/next')
+const { createVanillaExtractPlugin } = require('@vanilla-extract/next-plugin')
+
+const withVanillaExtract = createVanillaExtractPlugin()
 
 /**
  * @type {import('@nx/next/plugins/with-nx').WithNxOptions}
@@ -12,4 +15,10 @@ const nextConfig = {
   },
 }
 
-module.exports = withNx(nextConfig)
+const plugins = [
+  // Add more Next.js plugins to this list if needed.
+  withNx,
+  withVanillaExtract,
+]
+
+module.exports = composePlugins(...plugins)(nextConfig)
